Fix various configuration errors
The nginx configuration file was missing an events section, and other parts
that are required.  I initially excluded these because the example I was
copying from also excluded them...

Also, I tried to put a colon at the end of a function decorator in the flask
app.  Python did not like that one bit.