Add basics of all current API nodes
I'm not going to try and add the details for every single node, that should get
done as they are implemented/work starts on them.  These names are not the
standards that Chaim specified, but it's more organized this way.  If I have to
change everything back, then I will.