Add jquery to site
A minified jquery script has been added to the static content of the website.
To support this, a script tag has been added to the template, and will only be
added if the "jquery" variable is defined.  Also, a meta tag has been added to
the base template do define the charset as utf-8.