Check if user is authenticated before page loads
If a user navigating to the index page is already authenticated, we want them
to be redirected to their dashboard before the index page even loads.  For some
reason, I put the authentication-checking script inside the
wait-for-document-to-load block.  This has been fixed - it is now one of the
first things to happen.