Fix logout/login error
Because of the way the logout endpoint was implemented, when a user attempted
to log out, their session was never destroyed on the backend.  This would cause
an issue when the user attempts to re-log in.  The frontend would assume that
they are already logged in, and would not add their session cookie.